chore[5.13]: restructure code to support separated tests
