+++ src/utils/helpers.js
@@ -1,0 +1,9 @@
+const setStateTimeout = (setState, component, timeout) => {
+  setState(component)
+
+  setTimeout(() => {
+    setState('')
+  }, timeout)
+}
+
+export default { setStateTimeout }
